old-tests/webapps/WebSockets/tests/approved/interfaces/WebSocket/readyState/003.html

<!doctype html>
<title>WebSockets: delete readyState</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
test(function(t) {
  delete WebSocket.prototype.readyState;
  var ws = new WebSocket(SCHEME_AND_DOMAIN+'/');
  ws.close();
  delete ws.readyState;
  assert_equals(ws.readyState, ws.CLOSING);
});
</script>
