Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/websockets/interfaces/WebSocket/readyState/003-expected.txt


FAIL WebSockets: delete readyState assert_equals: delete ws.readyState expected 2 but got 3
